Restructure controller directory
