+++ src/client/commands/SubmitCmd.js
@@ -1,6 +1,6 @@
-import MorConfig from '../../controller/MorConfig.js'
-import { AlreadyExistsError, NotFoundError } from '../../controller/MorErrors.js'
-import MorUtils from '../../controller/MorUtils.js'
+import MorConfig from '../../controller/utils/MorConfig.js'
+import { AlreadyExistsError, NotFoundError } from '../../controller/utils/MorErrors.js'
+import MorUtils from '../../controller/utils/MorUtils.js'
 
 import { EmbedBuilder } from 'discord.js'
