+++ src/jobs/jobs/CreateBackup.js
@@ -1,4 +1,5 @@
-import MorConfig from '../../controller/MorConfig.js'
+import MorConfig from '../../controller/utils/MorConfig.js'
+
 import MorFacade from '../../controller/MorFacade.js'
 
 export default async function createBackup () {
